# Read the clock fewer times during message routing